feat: add trace-bound review path and exact cross-language trace hash
Build the headless-verifiable core of the review experience (Step 4 logic)
and make the cross-language trace hash exact.

Protocol: stream channel samples as IEEE-754 bit patterns (Vec<i64>) instead
of decimal JSON, and hash the canonical trace over little-endian bytes. Round-
tripping decimal text through two different float parsers (ryu vs Mono) is not
bit-guaranteed; transmitting bits makes the reconstructed hash identical on
both sides for real simulation data. Fixtures regenerated.

Unity review path (all trace-bound, no Transform/Rigidbody/VFX as truth source):
- EngineeringLayerCatalog: every layer declares its trace SourceIds, units, and
  a non-Unspecified evidence policy; uncertainty renders only when its channel
  exists.
- CameraDirector: the five-camera grammar (pad/chase/onboard/ground_tracking/
  inspection); switching is presentation-only and holds no clock.
- ReviewWorkbench: owns the playback clock, camera, and layer visibility; clean
  view hides all layers and restores them without touching time/camera/rate/play.
- CinematicExporter: export manifest carrying every provenance field
  (trace_sha256, protocol_version, mission_id, camera_id, window, preset,
  dimensions, frame_rate, unity_version, evidence_caveats).
- MissionRunner: drives one bridge run to an accepted immutable FlightTrace;
  a failed run never replaces the current one.

Tests: 31 EditMode (incl. layer traceability, camera grammar, export manifest,
and C#↔Rust hash parity against fixtures) + 5 PlayMode against the real Cargo
bridge — full review path (play, seek every event, all five cameras, layer
toggle, clean view, rerun with identical hash, export) plus cancellation. All
green on 6000.0.79f1. Workspace: 314 passed.

Deferred to interactive Unity work: HDRP scene/terrain, stage meshes, VFX plume,
Cinemachine rigs, UI Toolkit layout, and M3 FPS / unguided-review measurement.

+++ crates/ascent-visualizer-protocol/src/messages.rs
@@ -106,7 +106,11 @@ pub enum ServerMessage {
         run_id: String,
         channel: String,
         sequence: u32,
-        samples: Vec<f64>,
+        /// IEEE-754 f64 samples as little-endian bit patterns (reinterpreted as
+        /// i64). Transmitting bits — not decimal text — keeps the reconstructed
+        /// trace hash exact across languages, independent of each side's float
+        /// parser. Decode with `f64::from_bits(bits as u64)`.
+        samples: Vec<i64>,
     },
     TraceEvents {
         run_id: String,
